Seccion 12 finalizada

+++ src/modules/daybook/views/EntryView.vue
@@ -8,6 +8,12 @@
             </div>
             
             <div>
+                <input type="file" 
+                        @change="onSelectedImage"
+                        ref="imageSelector"
+                        v-show="false"
+                        accept="image/png, image/jpeg"
+                />
                 <button 
                     v-if="entry.id"
                     class="btn btn-danger mx-2"
@@ -16,8 +22,11 @@
                     Borrar
                     <i class="fa fa-trash-alt"></i>
                 </button>
-                <button class="btn btn-primary">
-                    Borrar
+                <button 
+                    class="btn btn-primary"
+                    @click="onSelectImage"
+                >
+                    Subir foto
                     <i class="fa fa-upload"></i>
                 </button>
             </div>
@@ -36,7 +45,14 @@
         </div>
         
         <img 
-            src="https://www.lifeder.com/wp-content/uploads/2020/11/paisaje-natural-y-antropico.jpg" 
+            v-if="entry.picture && !localImage"
+            :src="entry.picture" 
+            alt="entry-picture"
+            class="img-thumbnail"
+        >
+        <img 
+            v-if="localImage"
+            :src="localImage" 
             alt="entry-picture"
             class="img-thumbnail"
         >
@@ -52,8 +68,11 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
+import Swal from 'sweetalert2'
+
 
 import getDayMothYear from '../helpers/getDayMothYear'
+import uploadImage from '../helpers/uploadImage'
     
 export default {
     props: {
@@ -67,7 +86,9 @@ export default {
     },
     data(){
         return {
-            entry: null
+            entry: null,
+            localImage: null,
+            file: null
         }
     },
     computed: {
@@ -102,6 +123,17 @@ export default {
             this.entry = entry
         },
         async saveEntry(){
+            
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            
+            Swal.showLoading()
+            
+            const picture = await uploadImage( this.file )
+            this.entry.picture = picture
+            
             if( this.entry.id ){
                 await this.updateEntry(this.entry)
             } else {
@@ -110,10 +142,49 @@ export default {
                 this.$router.push({ name: 'entry', params: { id }})
                 
             }
+            
+            this.file = null
+            Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
         },
         async onDeleteEntry(){
-            await this.deleteEntry( this.entry.id )   
-            this.$router.push({ name: 'no-entry'})
+            
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Está seguro?',
+                text: 'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
+            
+            if( isConfirmed ){
+                new Swal({
+                    title: 'Espere por favor',
+                    allowOutsideClick: false
+                })
+                
+                Swal.showLoading()
+                await this.deleteEntry( this.entry.id )   
+                this.$router.push({ name: 'no-entry'})
+                Swal.fire('Eliminado', '', 'success')
+            }
+        },
+        onSelectedImage( event ){
+            const file = event.target.files[0] 
+            
+            if( !file ){
+                this.localImage = null
+                this.file = null
+                return
+            }
+            
+            this.file = file            
+            const fr = new FileReader()
+            
+            fr.onload = () => this.localImage = fr.result 
+            fr.readAsDataURL( file )
+            
+        },
+        onSelectImage(){
+            this.$refs.imageSelector.click()
         }
     },
     created(){
